renamed pax-cdi-sample8 -> pax-cdi-sample9 to avoid name clashing
