+++ itest/src/it/itest-standalone/pom.xml
@@ -139,7 +139,7 @@
         
         <dependency>
             <groupId>org.ops4j.pax.cdi.samples</groupId>
-            <artifactId>pax-cdi-sample8</artifactId>
+            <artifactId>pax-cdi-sample9</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
         </dependency>
