+++ pax-cdi-samples/pax-cdi-sample9/pom.xml
@@ -7,10 +7,10 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
     <groupId>org.ops4j.pax.cdi.samples</groupId>
-    <artifactId>pax-cdi-sample8</artifactId>
+    <artifactId>pax-cdi-sample9</artifactId>
     <packaging>bundle</packaging>
 
-    <name>OPS4J Pax CDI Sample8</name>
+    <name>OPS4J Pax CDI Sample9</name>
 
     <dependencies>
 
@@ -29,7 +29,7 @@
                 <artifactId>maven-bundle-plugin</artifactId>
                 <configuration>
                     <instructions>
-                        <Bundle-SymbolicName>org.ops4j.pax.cdi.sample8</Bundle-SymbolicName>
+                        <Bundle-SymbolicName>org.ops4j.pax.cdi.sample9</Bundle-SymbolicName>
                         <Require-Capability>
                             org.ops4j.pax.cdi.extension;
                             filter:="(&amp;(extension=pax-cdi-extension)(version&gt;=${version;==;${pax.cdi.osgi.version.clean}})(!(version&gt;=${version;=+;${pax.cdi.osgi.version.clean}})))",
